change int to long

+++ container.c
@@ -34,7 +34,7 @@ void runtest() {
     Vector *vec = new_vector();
     expect(__LINE__, 0, vec->len);
 
-    for (int i = 0; i < 100; i++)
+    for (long i = 0; i < 100; i++)
         vec_push(vec, (void *)i);
 
     expect(__LINE__, 100, vec->len);
